+++ src/main/java/com/example/communityapplication/service/ContentService.java
@@ -5,7 +5,7 @@ import com.example.communityapplication.model.*;
 import java.util.List;
 
 public interface ContentService {
-    public Content getById(Long contentId);
+    public Content getById(int contentId);
     public Content saveContent(ContentForm contentForm, User user);
     public List<ContentCard> getContentCardsByUser(User user);
 }
